move offset logic into scroll-into-view
