--- app/modifiers/scroll-to-element-with-offset.js
@@ -1,1 +1,0 @@
-export { default } from 'ember-scroll-modifiers/modifiers/scroll-to-element-with-offset';
