+++ docs/modifiers/scroll-into-view.md
@@ -1,19 +1,41 @@
 # scroll-into-view
 
-This modifier calls `scrollIntoView` on the modified element.
+This modifier scrolls to the associated element. By default it uses `scrollIntoView`, but if an offset is passed as an option it uses `scrollTo` and calculates the `options.top` attribute.
 
 
 ## When you should use this modifier
 
-You should use this modifier whenever you need to have an element scrolled into view on element insert.
+You should use this modifier whenever you need to have an element scrolled into view. If there is a element, such as a fixed header, passing in an `offset` will scroll to the element minus the `offset` value.
+
 
 
 ## Basic Usage
 
 `scroll-into-view` expects the named `shouldScroll` parameter and an optional `options` named parameter. See [scrollIntoView](https://developer.mozilla.org/en-US/docs/Web/API/Element/scrollIntoView) for the list of possible values and properties of `options`.
 
-```handlebars{data-execute=false}
-<div {{scroll-into-view shouldScroll=this.shouldScrollPromise options=true}}></div>
+
+```handlebars
+  <div {{scroll-into-view shouldScroll=this.shouldScroll options=(hash behavior="smooth")}}>
+    <button type="button" {{on "click" this.onScrollIntoView}}>
+      Trigger scroll-into-view on click
+    </button>
+  </div>
+```
+
+`shouldScroll` can be either a Boolean or a Promise that resolves to a truthy or falsy value. It does not handle a rejected Promise.
+
+
+### Usage with offset
+
+When passing in an offset, it will call `scrollTo`. See [scrollTo](https://developer.mozilla.org/en-US/docs/Web/API/Element/scrollTo) for the list of possible values and properties of `options`.
+
+```handlebars
+  <div {{scroll-into-view shouldScroll=this.shouldScrollWithOffset options=(hash offset=this.offset behavior="smooth")}}>
+    <input type="number" value={{this.offset}} {{on "change" this.onOffsetChange}}>
+    <button type="button" {{on "click" this.onScrollIntoViewWithOffset}}>
+      Trigger scroll-into-view with offset on click
+    </button>
+  </div>
 ```
 
 `shouldScroll` can be either a Boolean or a Promise that resolves to a truthy or falsy value. It does not handle a rejected Promise.
@@ -45,9 +67,19 @@ function('test scroll into view', (assert) => {
   ...
   assert.ok(this.mockHelperFunctions.scrollIntoViewCalledWith('[data-test-scroll-into-view-selector]'), 'element scrolled into view');
 });
+
+function('test scroll into view with offset', (assert) => {
+  ...
+  await render(
+    hbs`<div {{scroll-into-view shouldScroll=true options=(hash offset=25)}} data-test-scroll-into-view-selector></div>`
+  );
+  ...
+  assert.ok(this.mockHelperFunctions.scrollIntoViewCalledWith('[data-test-scroll-into-view-selector]', { behavior: 'smooth', top: 25, left: 0 }), 'scrolled to element');
+});
 ```
 
 
 ## Browser Support
 
 This feature is [supported](https://caniuse.com/?search=scrollIntoView) in the latest versions of every browser.
+This feature is [supported](https://caniuse.com/?search=scrollTo) in the latest versions of every browser.
