+++ tests/integration/modifiers/scroll-into-view-mock-test.js
@@ -46,5 +46,46 @@ module(
         )
       );
     });
+
+    test('scroll into view returns correct parameters with using offset', async function (assert) {
+      this.options = {
+        offset: 10,
+        behavior: 'auto',
+        left: 10,
+      };
+
+      await render(
+        hbs`<div style="height: 10px" {{scroll-into-view shouldScroll=true options=this.options}} data-test-scroll-into-view-selector></div>
+        <div style="height: 10px" data-test-scroll-into-view-not-modified></div>
+        <div style="height: 10px" {{scroll-into-view shouldScroll=false options=this.options}} data-test-scroll-into-view-not-scrolled></div>`
+      );
+
+      assert.ok(
+        this.mockHelperFunctions.scrollIntoViewCalledWith(
+          '[data-test-scroll-into-view-selector]',
+          this.options
+        ),
+        'scrollIntoViewCalledWith should return true for the given element selector'
+      );
+      assert.ok(
+        this.mockHelperFunctions.scrollIntoViewCalledWith(
+          find('[data-test-scroll-into-view-selector]'),
+          this.options
+        ),
+        'scrollIntoViewCalledWith should return true for the given element'
+      );
+      assert.notOk(
+        this.mockHelperFunctions.scrollIntoViewCalledWith(
+          '[data-test-scroll-into-view-not-modified]'
+        ),
+        "scrollIntoViewCalledWith should return false if the element doesn't have the modifier"
+      );
+      assert.notOk(
+        this.mockHelperFunctions.scrollIntoViewCalledWith(
+          '[data-test-scroll-into-view-not-scrolled]'
+        ),
+        'scrollIntoViewCalledWith should return false if the modifier was not triggered'
+      );
+    });
   }
 );
