+++ tests/integration/modifiers/scroll-into-view-test.js
@@ -97,4 +97,135 @@ module('Integration | Modifier | scroll-into-view', function (hooks) {
       'scrollIntoView was not called'
     );
   });
+
+  module('with offset', function (_hooks) {
+    _hooks.beforeEach(function () {
+      this.scrollToSpy = sandbox.spy(window, 'scrollTo');
+      this.getBoundingClientRectStub = sandbox.stub(
+        Element.prototype,
+        'getBoundingClientRect'
+      );
+      this.getBoundingClientRectStub.onCall(0).returns({ top: 100 });
+      this.getBoundingClientRectStub.onCall(1).returns({ top: 25 });
+    });
+
+    _hooks.afterEach(function () {
+      this.scrollToSpy = null;
+      sandbox.restore();
+    });
+
+    test('it renders and passes default options to scrollTo', async function (assert) {
+      this.options = {
+        offset: 50,
+      };
+
+      await render(
+        hbs`<div {{scroll-into-view shouldScroll=true options=this.options}}></div>`
+      );
+
+      assert.strictEqual(
+        this.scrollToSpy.args[0][0].behavior,
+        'auto',
+        'scrollTo was called with correct params'
+      );
+
+      assert.strictEqual(
+        this.scrollToSpy.args[0][0].left,
+        0,
+        'scrollTo was called with correct params'
+      );
+    });
+
+    test('it renders and passes behavior and left options to scrollTo', async function (assert) {
+      this.options = {
+        offset: 50,
+        behavior: 'smooth',
+        left: 10,
+      };
+
+      await render(
+        hbs`<div {{scroll-into-view shouldScroll=true options=this.options}}></div>`
+      );
+
+      assert.strictEqual(
+        this.scrollToSpy.args[0][0].behavior,
+        'smooth',
+        'scrollTo was called with correct params'
+      );
+
+      assert.strictEqual(
+        this.scrollToSpy.args[0][0].left,
+        10,
+        'scrollTo was called with correct params'
+      );
+    });
+
+    test('it renders and calculates correct top offset for scrollTo when offset is passed in', async function (assert) {
+      this.options = {
+        offset: 50,
+      };
+
+      await render(
+        hbs`<div id="test" {{scroll-into-view shouldScroll=true options=this.options}}></div>`
+      );
+
+      assert.deepEqual(
+        this.scrollToSpy.args[0][0],
+        {
+          behavior: 'auto',
+          left: 0,
+          top: 25,
+        },
+        'scrollTo was called with correct params'
+      );
+    });
+
+    test('it does not call scrollTo when shouldScroll is false', async function (assert) {
+      this.options = {
+        offset: 50,
+      };
+
+      await render(
+        hbs`<div {{scroll-into-view shouldScroll=false options=this.options}}></div>`
+      );
+
+      assert.notOk(this.scrollToSpy.called, 'scrollTo was not called');
+    });
+
+    test('it renders when shouldScroll resolves to true', async function (assert) {
+      this.options = {
+        offset: 50,
+      };
+
+      let resolvePromise;
+      this.shouldScroll = new Promise((resolve) => (resolvePromise = resolve));
+
+      await render(
+        hbs`<div {{scroll-into-view shouldScroll=this.shouldScroll options=this.options}}></div>`
+      );
+
+      assert.ok(this.scrollToSpy.notCalled, 'scrollTo was not called');
+
+      await resolvePromise(true);
+
+      assert.ok(this.scrollToSpy.called, 'scrollTo was called');
+    });
+
+    test('it does not render when shouldScroll resolves to false', async function (assert) {
+      this.options = {
+        offset: 50,
+      };
+
+      let resolvePromise;
+      this.shouldScroll = new Promise((resolve) => (resolvePromise = resolve));
+
+      await render(
+        hbs`<div {{scroll-into-view shouldScroll=this.shouldScroll options=this.options}}></div>`
+      );
+
+      await resolvePromise(false);
+
+      assert.ok(this.scrollToSpy.notCalled, 'scrollTo was not called');
+    });
+  });
 });
